웹 프로그래밍/django_crud/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-extensions==2.2.5
Faker==2.0.3
get==2019.4.13
idna==2.8
post==2019.4.13
public==2019.4.13
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.6
